analysis-requirements.txt
# Core data processing packages
pandas==2.1.4
numpy==1.24.4

# Visualization packages
matplotlib==3.7.4
seaborn==0.13.0
plotly==5.17.0

# Network analysis
networkx==3.2.1

# Scientific computing
scipy==1.11.4

# Database and SQL processing
duckdb==0.9.2
sqlparse==0.4.4
pyarrow==14.0.1

# Additional DuckDB-related packages for S3 and cloud storage
fsspec==2023.12.2
s3fs==2023.12.2

# File processing
openpyxl==3.1.2
xlrd==2.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Image processing
Pillow==10.1.0
pytesseract==0.3.10

# HTTP requests
requests==2.31.0

# Jupyter support (for notebook-style outputs)
ipython==8.18.1

# Additional data processing utilities
psutil==5.9.6
python-dateutil==2.8.2
pytz==2023.3
